+++ database_builder/bin/main.dart
@@ -1,16 +1,14 @@
+import 'package:database_builder/src/create_database.dart';
 import 'package:database_builder/src/transliterate.dart';
-// import 'package:database_builder/src/interlinear_text.dart';
-// import 'package:database_builder/src/transliterate.dart';
-// import 'package:database_builder/src/utils/interlinear_utils.dart';
 
 void main(List<String> arguments) {
   // printAllFilenames();
   // printAllParatextMarkers();
-  // createDatabase();
+  createDatabase();
 
   // transliterate();
   // createInterlinearText();
-  testTransliterator();
+  // testTransliterator();
 
   // compareColumns();
   // uniqueLanguageValues();